pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "marktypist"
version = "0.0.1"
authors = [
    # 替换成你自己的信息
    { name = "Your Name", email = "your.email@example.com" },
]
description = "一个功能强大的 Markdown 与 Typst 双向转换器"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Text Processing :: Markup",
]
dependencies = [
    "click ~= 8.1",               # 用于构建命令行接口。~=8.1 表示 >=8.1 且 <9.0
    "markdown-it-py ~= 3.0.0",    # Markdown 解析器
    "lark ~= 1.1.8",              # Typst 解析器生成工具
    "linkify-it-py ~= 2.0.2",  # <--- 新增这一行
]

[project.optional-dependencies]
# 开发环境依赖，包括测试工具
dev = [
    "pytest ~= 7.4",
]

[project.scripts]
# 这行代码是关键，它创建了 'marktypist' 命令行命令
marktypist = "marktypist.cli:cli"